+++ cpp/src/utilities/cython.cu
@@ -52,12 +52,12 @@ create_graph(raft::handle_t const& handle, graph_container_t const& graph_contai
     reinterpret_cast<vertex_t*>(graph_container.vertex_partition_offsets) +
       (graph_container.row_comm_size * graph_container.col_comm_size) + 1);
 
-  experimental::partition_t<int> partition(partition_offsets_vector,
-                                           graph_container.hypergraph_partitioned,
-                                           graph_container.row_comm_size,
-                                           graph_container.col_comm_size,
-                                           graph_container.row_comm_rank,
-                                           graph_container.col_comm_rank);
+  experimental::partition_t<vertex_t> partition(partition_offsets_vector,
+                                                graph_container.hypergraph_partitioned,
+                                                graph_container.row_comm_size,
+                                                graph_container.col_comm_size,
+                                                graph_container.row_comm_rank,
+                                                graph_container.col_comm_rank);
 
   return std::make_unique<experimental::graph_t<vertex_t, edge_t, weight_t, transposed, multi_gpu>>(
     handle,
@@ -66,7 +66,9 @@ create_graph(raft::handle_t const& handle, graph_container_t const& graph_contai
     static_cast<vertex_t>(graph_container.num_global_vertices),
     static_cast<edge_t>(graph_container.num_global_edges),
     graph_container.graph_props,
-    graph_container.sorted_by_degree,
+    // FIXME:  This currently fails if sorted_by_degree is true...
+    // graph_container.sorted_by_degree,
+    false,
     graph_container.do_expensive_check);
 }
 
@@ -270,30 +272,6 @@ void populate_graph_container_legacy(graph_container_t& graph_container,
 ////////////////////////////////////////////////////////////////////////////////
 
 namespace detail {
-template <typename graph_view_t, typename weight_t>
-std::pair<size_t, weight_t> call_louvain(raft::handle_t const& handle,
-                                         graph_view_t const& graph_view,
-                                         void* identifiers,
-                                         void* parts,
-                                         size_t max_level,
-                                         weight_t resolution)
-{
-  thrust::copy(  // rmm::exec_policy(handle.get_stream())->on(handle.get_stream()),
-    thrust::device,
-    thrust::make_counting_iterator(graph_view.get_local_vertex_first()),
-    thrust::make_counting_iterator(graph_view.get_local_vertex_last()),
-    reinterpret_cast<typename graph_view_t::vertex_type*>(identifiers));
-
-  return louvain(handle,
-                 graph_view,
-                 reinterpret_cast<typename graph_view_t::vertex_type*>(parts),
-                 max_level,
-                 static_cast<weight_t>(resolution));
-}
-
-}  // namespace detail
-
-namespace detail {
 
 // Final, fully-templatized call.
 template <bool transposed,
@@ -365,10 +343,10 @@ return_t call_function(raft::handle_t const& handle,
                        function_t function)
 {
   if (graph_container.weightType == numberTypeEnum::floatType) {
-    return call_function<transposed, return_t, function_t, float, transposed>(
+    return call_function<transposed, return_t, function_t, float, is_multi_gpu>(
       handle, graph_container, function);
   } else if (graph_container.weightType == numberTypeEnum::doubleType) {
-    return call_function<transposed, return_t, function_t, double, transposed>(
+    return call_function<transposed, return_t, function_t, double, is_multi_gpu>(
       handle, graph_container, function);
   } else {
     CUGRAPH_FAIL("weightType unsupported");
@@ -415,6 +393,11 @@ class louvain_functor {
   std::pair<size_t, weight_t> operator()(raft::handle_t const& handle,
                                          graph_view_t const& graph_view)
   {
+    thrust::copy(rmm::exec_policy(handle.get_stream())->on(handle.get_stream()),
+                 thrust::make_counting_iterator(graph_view.get_local_vertex_first()),
+                 thrust::make_counting_iterator(graph_view.get_local_vertex_last()),
+                 reinterpret_cast<typename graph_view_t::vertex_type*>(identifiers_));
+
     return cugraph::louvain(handle,
                             graph_view,
                             reinterpret_cast<typename graph_view_t::vertex_type*>(parts_),
@@ -466,7 +449,194 @@ std::pair<size_t, weight_t> call_louvain(raft::handle_t const& handle,
     handle, graph_container, functor);
 }
 
+// Wrapper for calling Pagerank through a graph container
+template <typename vertex_t, typename weight_t>
+void call_pagerank(raft::handle_t const& handle,
+                   graph_container_t const& graph_container,
+                   vertex_t* identifiers,
+                   weight_t* p_pagerank,
+                   vertex_t personalization_subset_size,
+                   vertex_t* personalization_subset,
+                   weight_t* personalization_values,
+                   double alpha,
+                   double tolerance,
+                   int64_t max_iter,
+                   bool has_guess)
+{
+  if (graph_container.graph_type == graphTypeEnum::GraphCSCViewFloat) {
+    pagerank(handle,
+             *(graph_container.graph_ptr_union.GraphCSCViewFloatPtr),
+             reinterpret_cast<float*>(p_pagerank),
+             static_cast<int32_t>(personalization_subset_size),
+             reinterpret_cast<int32_t*>(personalization_subset),
+             reinterpret_cast<float*>(personalization_values),
+             alpha,
+             tolerance,
+             max_iter,
+             has_guess);
+    graph_container.graph_ptr_union.GraphCSCViewFloatPtr->get_vertex_identifiers(
+      reinterpret_cast<int32_t*>(identifiers));
+  } else if (graph_container.graph_type == graphTypeEnum::GraphCSCViewDouble) {
+    pagerank(handle,
+             *(graph_container.graph_ptr_union.GraphCSCViewDoublePtr),
+             reinterpret_cast<double*>(p_pagerank),
+             static_cast<int32_t>(personalization_subset_size),
+             reinterpret_cast<int32_t*>(personalization_subset),
+             reinterpret_cast<double*>(personalization_values),
+             alpha,
+             tolerance,
+             max_iter,
+             has_guess);
+    graph_container.graph_ptr_union.GraphCSCViewDoublePtr->get_vertex_identifiers(
+      reinterpret_cast<int32_t*>(identifiers));
+  } else if (graph_container.graph_type == graphTypeEnum::graph_t) {
+    if (graph_container.edgeType == numberTypeEnum::int32Type) {
+      auto graph =
+        detail::create_graph<int32_t, int32_t, weight_t, true, true>(handle, graph_container);
+      cugraph::experimental::pagerank(handle,
+                                      graph->view(),
+                                      static_cast<weight_t*>(nullptr),
+                                      reinterpret_cast<int32_t*>(personalization_subset),
+                                      reinterpret_cast<weight_t*>(personalization_values),
+                                      static_cast<int32_t>(personalization_subset_size),
+                                      reinterpret_cast<weight_t*>(p_pagerank),
+                                      static_cast<weight_t>(alpha),
+                                      static_cast<weight_t>(tolerance),
+                                      max_iter,
+                                      has_guess,
+                                      false);
+    } else if (graph_container.edgeType == numberTypeEnum::int64Type) {
+      auto graph =
+        detail::create_graph<vertex_t, int64_t, weight_t, true, true>(handle, graph_container);
+      cugraph::experimental::pagerank(handle,
+                                      graph->view(),
+                                      static_cast<weight_t*>(nullptr),
+                                      reinterpret_cast<vertex_t*>(personalization_subset),
+                                      reinterpret_cast<weight_t*>(personalization_values),
+                                      static_cast<vertex_t>(personalization_subset_size),
+                                      reinterpret_cast<weight_t*>(p_pagerank),
+                                      static_cast<weight_t>(alpha),
+                                      static_cast<weight_t>(tolerance),
+                                      max_iter,
+                                      has_guess,
+                                      false);
+    } else {
+      CUGRAPH_FAIL("vertexType/edgeType combination unsupported");
+    }
+  }
+}
+
+// Wrapper for calling BFS through a graph container
+template <typename vertex_t, typename weight_t>
+void call_bfs(raft::handle_t const& handle,
+              graph_container_t const& graph_container,
+              vertex_t* identifiers,
+              vertex_t* distances,
+              vertex_t* predecessors,
+              double* sp_counters,
+              const vertex_t start_vertex,
+              bool directed)
+{
+  if (graph_container.graph_type == graphTypeEnum::GraphCSRViewFloat) {
+    graph_container.graph_ptr_union.GraphCSRViewFloatPtr->get_vertex_identifiers(
+      reinterpret_cast<int32_t*>(identifiers));
+    bfs(handle,
+        *(graph_container.graph_ptr_union.GraphCSRViewFloatPtr),
+        reinterpret_cast<int32_t*>(distances),
+        reinterpret_cast<int32_t*>(predecessors),
+        sp_counters,
+        static_cast<int32_t>(start_vertex),
+        directed);
+  } else if (graph_container.graph_type == graphTypeEnum::GraphCSRViewDouble) {
+    graph_container.graph_ptr_union.GraphCSRViewDoublePtr->get_vertex_identifiers(
+      reinterpret_cast<int32_t*>(identifiers));
+    bfs(handle,
+        *(graph_container.graph_ptr_union.GraphCSRViewDoublePtr),
+        reinterpret_cast<int32_t*>(distances),
+        reinterpret_cast<int32_t*>(predecessors),
+        sp_counters,
+        static_cast<int32_t>(start_vertex),
+        directed);
+  } else if (graph_container.graph_type == graphTypeEnum::graph_t) {
+    if (graph_container.edgeType == numberTypeEnum::int32Type) {
+      auto graph =
+        detail::create_graph<int32_t, int32_t, weight_t, false, true>(handle, graph_container);
+      cugraph::experimental::bfs(handle,
+                                 graph->view(),
+                                 reinterpret_cast<int32_t*>(distances),
+                                 reinterpret_cast<int32_t*>(predecessors),
+                                 static_cast<int32_t>(start_vertex));
+    } else if (graph_container.edgeType == numberTypeEnum::int64Type) {
+      auto graph =
+        detail::create_graph<vertex_t, int64_t, weight_t, false, true>(handle, graph_container);
+      cugraph::experimental::bfs(handle,
+                                 graph->view(),
+                                 reinterpret_cast<vertex_t*>(distances),
+                                 reinterpret_cast<vertex_t*>(predecessors),
+                                 static_cast<vertex_t>(start_vertex));
+    } else {
+      CUGRAPH_FAIL("vertexType/edgeType combination unsupported");
+    }
+  }
+}
+
+// Wrapper for calling SSSP through a graph container
+template <typename vertex_t, typename weight_t>
+void call_sssp(raft::handle_t const& handle,
+               graph_container_t const& graph_container,
+               vertex_t* identifiers,
+               weight_t* distances,
+               vertex_t* predecessors,
+               const vertex_t source_vertex)
+{
+  if (graph_container.graph_type == graphTypeEnum::GraphCSRViewFloat) {
+    graph_container.graph_ptr_union.GraphCSRViewFloatPtr->get_vertex_identifiers(
+      reinterpret_cast<int32_t*>(identifiers));
+    sssp(  // handle, TODO: clarify: no raft_handle_t? why?
+      *(graph_container.graph_ptr_union.GraphCSRViewFloatPtr),
+      reinterpret_cast<float*>(distances),
+      reinterpret_cast<int32_t*>(predecessors),
+      static_cast<int32_t>(source_vertex));
+  } else if (graph_container.graph_type == graphTypeEnum::GraphCSRViewDouble) {
+    graph_container.graph_ptr_union.GraphCSRViewDoublePtr->get_vertex_identifiers(
+      reinterpret_cast<int32_t*>(identifiers));
+    sssp(  // handle, TODO: clarify: no raft_handle_t? why?
+      *(graph_container.graph_ptr_union.GraphCSRViewDoublePtr),
+      reinterpret_cast<double*>(distances),
+      reinterpret_cast<int32_t*>(predecessors),
+      static_cast<int32_t>(source_vertex));
+  } else if (graph_container.graph_type == graphTypeEnum::graph_t) {
+    if (graph_container.edgeType == numberTypeEnum::int32Type) {
+      auto graph =
+        detail::create_graph<int32_t, int32_t, weight_t, false, true>(handle, graph_container);
+      cugraph::experimental::sssp(handle,
+                                  graph->view(),
+                                  reinterpret_cast<weight_t*>(distances),
+                                  reinterpret_cast<int32_t*>(predecessors),
+                                  static_cast<int32_t>(source_vertex));
+    } else if (graph_container.edgeType == numberTypeEnum::int64Type) {
+      auto graph =
+        detail::create_graph<vertex_t, int64_t, weight_t, false, true>(handle, graph_container);
+      cugraph::experimental::sssp(handle,
+                                  graph->view(),
+                                  reinterpret_cast<weight_t*>(distances),
+                                  reinterpret_cast<vertex_t*>(predecessors),
+                                  static_cast<vertex_t>(source_vertex));
+    } else {
+      CUGRAPH_FAIL("vertexType/edgeType combination unsupported");
+    }
+  }
+}
+
+// Helper for setting up subcommunicators
+void init_subcomms(raft::handle_t& handle, size_t row_comm_size)
+{
+  partition_2d::subcomm_factory_t<partition_2d::key_naming_t, int> subcomm_factory(handle,
+                                                                                   row_comm_size);
+}
+
 // Explicit instantiations
+
 template std::pair<size_t, float> call_louvain(raft::handle_t const& handle,
                                                graph_container_t const& graph_container,
                                                void* identifiers,
@@ -481,12 +651,117 @@ template std::pair<size_t, double> call_louvain(raft::handle_t const& handle,
                                                 size_t max_level,
                                                 double resolution);
 
-// Helper for setting up subcommunicators
-void init_subcomms(raft::handle_t& handle, size_t row_comm_size)
-{
-  partition_2d::subcomm_factory_t<partition_2d::key_naming_t, int> subcomm_factory(handle,
-                                                                                   row_comm_size);
-}
+template void call_pagerank(raft::handle_t const& handle,
+                            graph_container_t const& graph_container,
+                            int* identifiers,
+                            float* p_pagerank,
+                            int32_t personalization_subset_size,
+                            int32_t* personalization_subset,
+                            float* personalization_values,
+                            double alpha,
+                            double tolerance,
+                            int64_t max_iter,
+                            bool has_guess);
+
+template void call_pagerank(raft::handle_t const& handle,
+                            graph_container_t const& graph_container,
+                            int* identifiers,
+                            double* p_pagerank,
+                            int32_t personalization_subset_size,
+                            int32_t* personalization_subset,
+                            double* personalization_values,
+                            double alpha,
+                            double tolerance,
+                            int64_t max_iter,
+                            bool has_guess);
+
+template void call_pagerank(raft::handle_t const& handle,
+                            graph_container_t const& graph_container,
+                            int64_t* identifiers,
+                            float* p_pagerank,
+                            int64_t personalization_subset_size,
+                            int64_t* personalization_subset,
+                            float* personalization_values,
+                            double alpha,
+                            double tolerance,
+                            int64_t max_iter,
+                            bool has_guess);
+
+template void call_pagerank(raft::handle_t const& handle,
+                            graph_container_t const& graph_container,
+                            int64_t* identifiers,
+                            double* p_pagerank,
+                            int64_t personalization_subset_size,
+                            int64_t* personalization_subset,
+                            double* personalization_values,
+                            double alpha,
+                            double tolerance,
+                            int64_t max_iter,
+                            bool has_guess);
+
+template void call_bfs<int32_t, float>(raft::handle_t const& handle,
+                                       graph_container_t const& graph_container,
+                                       int32_t* identifiers,
+                                       int32_t* distances,
+                                       int32_t* predecessors,
+                                       double* sp_counters,
+                                       const int32_t start_vertex,
+                                       bool directed);
+
+template void call_bfs<int32_t, double>(raft::handle_t const& handle,
+                                        graph_container_t const& graph_container,
+                                        int32_t* identifiers,
+                                        int32_t* distances,
+                                        int32_t* predecessors,
+                                        double* sp_counters,
+                                        const int32_t start_vertex,
+                                        bool directed);
+
+template void call_bfs<int64_t, float>(raft::handle_t const& handle,
+                                       graph_container_t const& graph_container,
+                                       int64_t* identifiers,
+                                       int64_t* distances,
+                                       int64_t* predecessors,
+                                       double* sp_counters,
+                                       const int64_t start_vertex,
+                                       bool directed);
+
+template void call_bfs<int64_t, double>(raft::handle_t const& handle,
+                                        graph_container_t const& graph_container,
+                                        int64_t* identifiers,
+                                        int64_t* distances,
+                                        int64_t* predecessors,
+                                        double* sp_counters,
+                                        const int64_t start_vertex,
+                                        bool directed);
+
+template void call_sssp(raft::handle_t const& handle,
+                        graph_container_t const& graph_container,
+                        int32_t* identifiers,
+                        float* distances,
+                        int32_t* predecessors,
+                        const int32_t source_vertex);
+
+template void call_sssp(raft::handle_t const& handle,
+                        graph_container_t const& graph_container,
+                        int32_t* identifiers,
+                        double* distances,
+                        int32_t* predecessors,
+                        const int32_t source_vertex);
+
+template void call_sssp(raft::handle_t const& handle,
+                        graph_container_t const& graph_container,
+                        int64_t* identifiers,
+                        float* distances,
+                        int64_t* predecessors,
+                        const int64_t source_vertex);
+
+template void call_sssp(raft::handle_t const& handle,
+                        graph_container_t const& graph_container,
+                        int64_t* identifiers,
+                        double* distances,
+                        int64_t* predecessors,
+                        const int64_t source_vertex);
 
 }  // namespace cython
 }  // namespace cugraph