+++ python/cugraph/layout/force_atlas2_wrapper.pyx
@@ -127,7 +127,7 @@ def force_atlas2(input_graph,
                 <bool> verbose,
                 <GraphBasedDimRedCallback*>callback_ptr)
 
-        pos_df = cudf.DataFrame.from_gpu_matrix(pos, columns=['x', 'y'])
+        pos_df = cudf.DataFrame(pos, columns=['x', 'y'])
         df['x'] = pos_df['x']
         df['y'] = pos_df['y']
     else:
@@ -159,7 +159,7 @@ def force_atlas2(input_graph,
                 <bool> verbose,
                 <GraphBasedDimRedCallback*>callback_ptr)
 
-        pos_df = cudf.DataFrame.from_gpu_matrix(pos, columns=['x', 'y'])
+        pos_df = cudf.DataFrame(pos, columns=['x', 'y'])
         df['x'] = pos_df['x']
         df['y'] = pos_df['y']